requirements.txt
pillow==11.0.0
tensorflow==2.18.0
streamlit==1.44.0